Extracting homologous UCEs on chromosome clusters
We chose a suite of chromosomes based on the clustering pattern obtained through the previous step - MMDS based on the UCEs mapped on all the chromosomes of our 10 sample taxa. These chromosomes are very similar in their UCE content, however they are not identical. Therefore the first step is to remove UCEs that are not common to all the 10 sample chromosomes to obtain a set of homologous UCEs. In the next step we extract physical location of these homologous UCEs and carryout a PCA to quantify difference between these taxa. We also size correct the UCE positions by dividing them with the chromosome size to remove one level of allometry.